fix(boot): 增强HardFault调试信息并修复分布式总线格式错误
在HardFault_Handler中新增LR/MSP/PSP寄存器、异常栈帧、CFSR/HFSR/MMFAR/BFAR寄存器以及故障任务ID的打印，便于定位崩溃原因。同时修正了分布式总线beacon消息中challenge字段的格式化参数类型。

+++ net/distributed_bus.hpp
@@ -300,7 +300,7 @@ public:
         snprintf(beacon_payload, sizeof(beacon_payload),
             "{\"event\":\"beacon\",\"device_id\":\"%s\","
             "\"cap\":[\"display\",\"touch\"],\"seq\":\"%lu\",\"auth\":\"%s\"}",
-            (unsigned long)challenge, (unsigned long)seq, auth_hex);
+            challenge, (unsigned long)seq, auth_hex);
 
         lwip_sendto(udp_socket_, beacon_payload,
                     strlen(beacon_payload), 0,
